Rudimentary BackupViewer implementation
